Working On Fines

+++ Api/Library.Api/Library.Application/Features/Fines/Dtos/FineDto.cs
@@ -3,4 +3,4 @@
 namespace Library.Application.Features.Fines.Dtos;
 
 public sealed record FineDto(Guid Id,UserInfoDto UserInfoDto,BorrowingRecordDto BorrowingRecordDto,Guid BorrowingRecordId,
-    int NumberOfLateDays,decimal FineAmount,PaymentStatus PaymentStatus);
+    double NumberOfLateDays,double FineAmount,PaymentStatus PaymentStatus);
